Don't write the file if we don't need to o3o

+++ api/v1/build_playlists.php
@@ -24,6 +24,10 @@
 				$current_list .= "\n$file";
 			else $current_list = iterate_dir(MEDIA_DIR . $file, $current_list, $exclusions);
 		}
+        $existing = file_get_contents(MEDIA_DIR . $name)
+        if strcmp($existing, $current_list) == 0 {
+            return
+        }
 		file_put_contents(MEDIA_DIR . $name, $current_list);
 	}
